# api-definitions/openapi/skedular/organization_v1.yaml
openapi: 3.1.1
info:
  title: Organization API v1
  description: "Organization REST API"
  version: 1.0.0

servers:
  - url: https://api.skedular.app
    description: production
  - url: https://apistaging.skedular.app
    description: staging
  - url: http://localhost:9000
    description: local gateway
  - url: http://localhost:10200
    description: local

paths:
  /v1/organization/version:
    get:
      tags:
        - v1
        - organization
      summary: return API version
      operationId: getVersion

      responses:
        "200":
          description: the version of the API
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Version"

        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /v1/organization/{organizationId}/republish:
    put:
      tags:
        - v1
        - organization
      summary: republish organization
      operationId: republish
      parameters:
        - in: path
          name: organizationId
          schema:
            type: string
          required: true

      responses:
        "200":
          description: the status of organization republishing

        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /v1/organization/republish-all:
    put:
      tags:
        - v1
        - organization
      summary: republish all organizations
      operationId: republishAll

      responses:
        "200":
          description: the status of organization republishing

        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /v1/organization/azure-tenant-admin-consent-url:
    get:
      tags:
        - v1
        - organization
        - azure
        - tenant
      summary: generate an admin consent Url for the given tenant
      operationId: azureTenantAdminConsentUrl
      responses:
        "302":
          description: redirect status response code

  /v1/organization/onboard-azure-tenant:
    get:
      tags:
        - v1
        - organization
        - azure
        - tenant
      summary: OnBoard a tenant
      operationId: onboardAzureTenant
      parameters:
        - in: query
          name: tenant
          schema:
            type: string
          required: true
        - in: query
          name: admin_consent
          schema:
            type: boolean
          required: true
        - in: query
          name: state
          schema:
            type: string
          required: true
        - in: query
          name: error
          schema:
            type: string
        - in: query
          name: error_description
          schema:
            type: string
      responses:
        "200":
          description: onboarding response

  /v1/organization/acs:
    post:
      tags:
        - v1
        - organization
        - sso
      summary: sso acs
      operationId: SsoAcs

      responses:
        "200":
          description: sso acs

        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /v1/organization/add-payment-method:
    get:
      tags:
        - v1
        - organization
        - payment-method
      summary: add payment method
      operationId: addPaymentMethod
      parameters:
        - name: setup_intent
          in: query
          required: true
          schema:
            type: string
        - name: setup_intent_client_secret
          in: query
          required: true
          schema:
            type: string
        - name: redirect_status
          in: query
          required: true
          schema:
            type: string

      responses:
        "200":
          description: the readiness status

        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /v1/organization/organization-stripe-connect-account/refresh-onboarding-url:
    get:
      tags:
        - v1
        - organization
        - organizationStripeConnectAccounts
        - onboarding
      summary: return OrganizationStripeConnectAccount onboarding refresh URL
      operationId: refreshOrganizationStripeConnectAccountOnboarding
      parameters:
        - in: query
          name: code
          schema:
            type: string
          required: true

      responses:
        "200":
          description: should never be returned
        "302":
          description: redirect to OrganizationStripeConnectAccounts new onboarding URL

        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /v1/organization/stripe/platform/account/webhook:
    post:
      tags:
        - v1
        - organization
        - stripe
        - webhook
        - platform
      summary: Stripe Platform Account Webhook
      operationId: processStripePlatformAccountEvent
      parameters:
        - in: header
          name: Stripe-Signature
          schema:
            type: string
          required: false
          description: Stripe webhook signature

      responses:
        "200":
          description: the status of processing the Stripe Platform Account event

        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /v1/organization/stripe/connect/account/webhook:
    post:
      tags:
        - v1
        - organization
        - stripe
        - webhook
        - connect
      summary: Stripe Connect Account Webhook
      operationId: processStripeConnectAccountEvent
      parameters:
        - in: header
          name: Stripe-Signature
          schema:
            type: string
          required: false
          description: Stripe webhook signature

      responses:
        "200":
          description: the status of processing the Stripe Connect Account event

        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Error code
        message:
          type: string
          description: Error message

    Version:
      type: object
      required:
        - major
        - minor
        - build
        - revision
      properties:
        major:
          type: integer
          format: int32
          description: major version
        minor:
          type: integer
          format: int32
          description: minor version
        build:
          type: integer
          format: int32
          description: build number
        revision:
          type: integer
          format: int32
          description: revision